+++ config/wagmi/.eslintrc.mjs
@@ -1,0 +1,1 @@
+export { default } from '@sushiswap/eslint-config'