+++ packages/wagmi/future/hooks/pools/actions/getAllPoolsCodeMap.ts
@@ -9,7 +9,23 @@ export const getAllPoolsCodeMap = async ({ currencyA, currencyB, chainId }: Omit
     return nullPoolCodeMap
   }
   const dataFetcher = DataFetcher.onChain(chainId)
-  const liquidityProviders = [LiquidityProviders.SushiSwap, LiquidityProviders.Trident]
+  const liquidityProviders = [
+    LiquidityProviders.SushiSwapV2,
+    LiquidityProviders.Trident,
+    LiquidityProviders.UniswapV2,
+    LiquidityProviders.QuickSwap,
+    LiquidityProviders.ApeSwap,
+    LiquidityProviders.PancakeSwap,
+    LiquidityProviders.TraderJoe,
+    LiquidityProviders.Dfyn,
+    LiquidityProviders.Elk,
+    LiquidityProviders.JetSwap,
+    LiquidityProviders.SpookySwap,
+    LiquidityProviders.NetSwap,
+    LiquidityProviders.HoneySwap,
+    LiquidityProviders.UbeSwap,
+    LiquidityProviders.Biswap,
+  ]
   if (isRouteProcessor3ChainId(chainId)) {
     liquidityProviders.push(LiquidityProviders.SushiSwapV3)
   }