+++ packages/amm/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sushiswap/amm
 
+## 0.1.3
+
+### Patch Changes
+
+- 074ec9ebc: initial publish chunk
+- Updated dependencies [074ec9ebc]
+  - @sushiswap/trident@1.0.1
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @sushiswap/trident@1.0.0
+
 ## 0.1.1
 
 ### Patch Changes